ASoC:SOF:get spcm from FE dai link during BE dai link prepare
BE dai link prepare() needs to setup hw_params again only
if the internal flag (hw_params_upon_resume) is set
for the pcm stream.

But since BE dai links themselves don't have a pcm, we need
to use the FE dai link that is connected to the BE dai link
to check for the flag.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Signed-off-by: Rander Wang <rander.wang@linux.intel.com>