ASoC: SOF: assign link DMA channel at run-time
The recommended HDA HW programming sequence for setting
the DMA format requires that the link DMA and host DMA
channels be coupled before setting the format. This
change means that host DMA or link DMA channels be
reserved even if only one is used.

Statically assigned link DMA channels would mean that
all the corresponding host DMA channels will need to be
reserved, leaving only a few channels available at run-time.
So, the suggestion here is to switch to dynamically assigning
both host DMA channels and link DMA channels are run-time.

The host DMA channel is assigned when the pcm
is opened as before. While choosing the link DMA channel,
if the host DMA channel corresponding to the link DMA channel
is already taken, the proposed method checks to make
sure that the BE is connected to the FE that has been assigned
this host DMA channel. For hostless streams, the host DMA channel is
marked as opened to ensure it isn't assigned to another stream.
After assigning the link DMA channel, a DAI_CONFIG IPC is sent to
the FW to set the chosen link DMA channel.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>