ASoC: SOF: hda: set autosuspend delay for hda bus device
Set the autosuspend delay to 2s for hda bus device.
This is needed to prevent the codec driver from entering
runtime PM during stress tests involving back-to-back
stream open/close.

The legacy driver forbids autosuspend for some devices
where a non-zero autosuspend delay causes pops/clicks.
This needs to be addressed in the SOF driver when those
devices are supported.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>