ASoC: SOF: intel: get dai index from dai name
The dai index comes from topology. However, we don't have a good
mechanism to get the topology data in the hw_params or hw_free function.
The patch assume that there is only one BE dai on each soundwire bus.
And we should use the ALH port with the same index as soundwire bus.
e.g.
We should use ALH 1 if the codec is on soundwire bus 1.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>