soundwire: intel: stream deprepare on TRIGGER_STOP/SUSPEND/PAUSE_PUSH
ALSA does not have a deprepare step, and calling
sdw_deprepare_stream() in the .hw_free step raises all kinds of
issues, e.g. with aplay not resuming after system suspend since the
stream is already configured.

It's just better to call sdw_deprepare_stream() for all transfer
stops, and require that the .prepare() callback be used to reconfigure
the stream.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>