soundwire: stream: fix state machine and add state checks
The state machine is incorrect, we need to be able to go from
STREAM_DEPREPARED to STREAM_PREPARED, e.g. to deal with underflows or
suspend/resume, where the .prepare() callback is called after a
.trigger (STOP/SUSPEND).

Add checks for prepare, enable, disable, deprepared to make sure the
client is notified of illegal state machine transitions.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>