soundwire: intel: re-initialize ALH PDIs during suspend/resume transition
The .prepare() callback is invoked for underflows or during the system
resume transition. In the latter case, the context for the ALH PDIs is
lost, and the DSP is not initialized properly either.

Keep track of the TRIGGER_SUSPEND transition with a boolean flag, and
use it during the .prepare step to check if the ALH PDIs need to be
reconfigured.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>