soundwire: bus: implement initialization_complete signaling
init_completion() is called when the Slave device becomes unattached,
as done with enumeration_complete.

The difference with the enumeration_complete case is that complete()
is signaled when the Slave device is fully initialized after the
.update_status() callback is called.

A Slave device driver can decide to wait on either of the two
complete() cases, depending on its initialization code and
requirements.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>