ASoC: SOF: Intel: fix pop noise when stop playback on sdw platforms
Now the trigger sequence is set to SND_SOC_DPCM_TRIGGER_POST for
SOF. This means FE will be stopped before BE, so BE will consume
invalid data and this generates huge pop noise. Now set trigger
sequence to SND_SOC_DPCM_TRIGGER_PRE to let FE to be stopped after
BE to avoid this issue.

Fully tested on Comet Lake for a few cycles.

Signed-off-by: randerwang <rander.wang@linux.intel.com>