soundwire: stream: enforce order when adding devices to stream
The existing code allows for a Master or the peripheral devices to add
a runtime for a stream in any order. In practice, because of the order
imposed by ASoC-DPCM, the runtime for the master is *always* added by
the Slave driver. This seems to be a major layering violation.

Now that the ASoC-DPCM order deals with the cpu DAIs first, this patch
enforces a strict order where the master runtime has to be added
first, with no interaction or interference between codec and master
drivers. As an added benefit this greatly simplifies error handling
and allows for upcoming updates where the hw_params() function can be
called multiple times - currently a clear incompatibility between the
ALSA and SoundWire/stream designs.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>