ASoC: intel: sof_sdw: Use a fixed DAI link id for AMP
Currently, we assign SoundWire DAI link id according to the order in
snd_soc_acpi_link_adr. It will varies between devices and we have to
modify topologies for the different DAI link ids. Fix a fixed DAI link id
can reduce the possibility of editing topologies.
This patch has the assumption that the codec order in snd_soc_acpi_link_adr
is always headset -> amp -> mic. We have to follow the order when we
add new link adrs in the future.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>