ASoC: SOF: add direction to snd_sof_widget
A playback widget will connect to a capture widget in the echo reference
topology.
For example, demux.1.2 is connected to host.7.1

        host.1.1 --> demux.1.2 --> SSP.2.1
                        \
                         \
             SSP.8.1 --> host.7.1

When we start the playback PCM, we only want to setup/prepare the
widgets that belong to the playback pcm, and same as the capture PCM.

Adding direction in snd_sof_widget allows driver knows what is the
direction of a widget.

The direction can be none that implies that is a dummy widget for
firmwire and we should not send any IPC for the widget.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>