ASoC: Intel: sof_sdw: allow multi dai links on one sdw link
Currently, we will create a dai link for each matched acpi link adr
and construct the dai link name by direction and sdw link id.
It means that we will have exactly the same link name if two dai links
are on the same sdw link with the same direction.

To solve the issue and keep backward compatible this patch appends
codec dai name to a duplicated dai link name.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>