ASoC: Intel: append codec type to dai link name
The existing sdw_sof machine driver construct the dai link name by
direction and sdw link id. It means that we will have exactly the same
dai link name if two dai links are on the same sdw link with the same
direction.
The new Realtek codec has two SoundWire adrs for jack and DMIC
functions and it is treated as different codecs. To create two dai links
for jack and DMIC, we need to create different dai link names.
This patch suggests to append codec type to the second dai link name.
That wont impact the existing topology with this way.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>