ASoC: Intel: sof_sdw: add dai info
The existing code create a dailink for a codec. However, we may need
multi dailinks for a codec. This commit adds a new struct in
sof_sdw_codec_info{} to store the dai info of a codec.
We assume that we will create at most 3 dailinks for a codec since there
are at most 3 types of dais and each dai type can share a dailink.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>