ASoC: Intel: sof_sdw: use predefine dailink id
Currently, we assign dailink ids in order, and shift with codec type.
The purpose is to have consistent dailink ids for topologies.
This can be simplified if we have a predefined dailink id in
sof_sdw_dai_info.
With this change, we no longer need to check the adr order described in a
snd_soc_acpi_link_adr array.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>