ASoC: SOF: ipc4-topology: Modify the output format selection logic
Modify the output format selection when there are multiple output
formats available to choose the one that matches the reference params.
The reference params depend on the type of module. In the case of
processing modules, the reference params is based on the selected input
audio format.

The only special case is the copier module. In the case of
a module copier, the reference params is based on the selected input
params. In the case of host copier, the reference params should be
based on the input audio format for playback and the FE hw_params for
capture. In the case DAI copier, the reference params should be based on
the input audio format for capture and the FE hw_params for playback.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>