soundwire: add lane_mask in sdw_port_config
The DisCo table will only provide the lane mapping if the lane is
connected. The lane_maps[i] will be 0 by default and we can't tell if
a lane is connected to lane 0 or not connected.
Besides, codec may have restricting lane for ports. For example,
lane 0 can only be used by port 1.
Add a lane_mask to allow codec driver define the available lanes
for the port.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>