Soundwire: generic_bandwidth_allocation: select higher clk_values
Although we check if (curr_dr_freq <= bus->params.bandwidth) in the
first loop to get a suitable clk_values, sdw_select_row_col() could
return error because the bandwidth of control bits is not calculated
int the first loop. Reselect a higher clock value if the current clock
is not the maximum value.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>