soundwire: stream: skip m_rt which is not in preparing stage
sdw_compute_group_params() will count payload bandwidth and it count all
m_rt in the bus. The m_rt will be added to the bus by
sdw_stream_add_master(). If 2 streams are opened simultaneously, the
m_rt of the second stream will be counted when
sdw_compute_group_params() is counting the payload bandwidth of the
first stream. Which is incorrect.
Set PREPARED state earlier so that sdw_compute_group_params() can skip
the m_rt that is not in the preparing stage yet.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>