ASoC: soc_sdw_utils: add get_sdw_codec_name
We return the preset codec_info->codec_name in asoc_sdw_get_codec_name.
However, we always need to get the SoundWire codec name if we want to
find the SoundWire device. Add get_sdw_codec_name() for the case that
need to get the SoundWire codec name no matter codec_info->codec_name is
set or not.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>