ASoC/soundwire: send multi sections in one BPT stream
Currently we send a BRA message with a start address with continuous
registers in a BPT stream. However, a codec may need to write different
register sections shortly. It makes more sense to send different
register sections in a BPT stream and no need to close/open the BPT
stream repeatedly.

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>